refactor: move BeginMissionSequence class to new begin_mission_sequence.py
