+++ gmatpyplus/command/begin_mission_sequence.py
@@ -1,0 +1,13 @@
+import gmatpyplus as gp
+from command.gmat_command import GmatCommand
+
+
+class BeginMissionSequence(GmatCommand):
+    def __init__(self):
+        super().__init__('BeginMissionSequence', 'BeginMissionSequenceCommand')
+
+        self.SetSolarSystem()
+        self.SetObjectMap(gp.Moderator().GetConfiguredObjectMap())
+        self.SetGlobalObjectMap(gp.Sandbox().GetGlobalObjectMap())
+
+        self.Initialize()
